backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
google-genai==0.3.0
tweepy==4.14.0
apscheduler==3.10.4
python-dotenv==1.0.0
pydantic==2.6.0
httpx==0.26.0
python-multipart==0.0.6
requests==2.31.0
Pillow==10.2.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
email-validator==2.1.0
pytest==7.4.3
pytest-mock==3.12.0
